requirements.txt
Django==4.1.2
djangorestframework==3.14.0
djangorestframework-jsonapi==6.0.0
djangorestframework-jsonapi["django-filter"]==22.1
Pillow==9.2.0